requirements.txt
dj-database-url==0.4.1
Django==1.10
django-registration==2.2
gunicorn==19.6.0
psycopg2==2.6.1
OpenTok>=2.4
whitenoise==3.2
